requirements.txt
Flask==3.0.0
psutil==5.9.6
Werkzeug==3.0.1
PyYAML==6.0.1
requests==2.31.0
